[build] Move development tools into a new package thinp_dev
