+++ src/thinp_dev/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "thinp_dev"
+description = "Development tools for metadata derived from dm-persistent-data"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+clap = "3.0"
+fixedbitset = "0.4"
+num_cpus = "1.13"
+rand = "0.8"
+termion = "1.5"
+thinp = { path = "../thinp" }
+tui = "0.14"
