+++ src/thinp_dev/src/bin/thin_explore.rs
