+++ src/thinp_dev/src/cache/mod.rs
@@ -1,0 +1,1 @@
+pub mod metadata_generator;
